% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConversationAlign-package.R
\docType{package}
\name{ConversationAlign-package}
\alias{ConversationAlign}
\alias{ConversationAlign-package}
\title{ConversationAlign: ConversationAlign}
\description{
Reads conversation transcripts into R, cleans and formats them, and yokes user-specified psycholinguistic database values to each word. ConversationAlign computes alignment indices between two interlocutors across each transcript for up to 30 different psycholinguistic and affective lexical dimensions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Reilly-ConceptsCognitionLab/ConversationAlign}
}

}
\author{
\strong{Maintainer}: Jamie Reilly \email{jamie_reilly@temple.edu} (\href{https://orcid.org/0000-0002-0891-438X}{ORCID})

Authors:
\itemize{
  \item Virginia Ulichney
  \item Ben Sacks
}

Other contributors:
\itemize{
  \item Sarah Weinstein [contributor]
  \item Chelsea Helion [contributor]
  \item Gus Cooney [contributor]
}

}
\keyword{internal}
