% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_corpus.R
\name{describe_corpus}
\alias{describe_corpus}
\title{describe_corpus}
\usage{
describe_corpus(df, granularity = "none", export_table = TRUE)
}
\arguments{
\item{granularity}{provides analytics by person or by conversation, default is granularity="conversation"}

\item{df_aligned}{takes dataframe produced from the df_prep() function}
}
\value{
dataframe with summary analytics for a conversation corpus
}
\description{
Produces a table of corpus analytics including numbers of complete observations at each step, word counts, lexical diversity (e.g., TTR), stopword ratios, etc. Granularity of the summary statistics are guided by the user (e.g., by conversation, by conversation and speaker, collapsed all)
}
