% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_dyads.R
\name{inspect_dyads}
\alias{inspect_dyads}
\title{inspect_dyads}
\usage{
inspect_dyads(aligned_ts_df, dimensions = "all", include_plot = FALSE)
}
\arguments{
\item{aligned_ts_df}{aligned dataframe ported from the align_dyads() step}

\item{dimensions}{character vector of aligned dimensions to inspect. default selects all.}
}
\description{
Creates a grouped dataframe of word analytics by speaker and dyad, as well as a faceted graph showing each dyadic time series that is saved to the local disc.
}
