% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MaronGross_2013}
\alias{MaronGross_2013}
\title{Sample Dyadic Interview Transcript: Marc Maron and Terry Gross Radio Interview 2013}
\format{
## "MaronGross_2013"
A data.frame with 546 obs, 2 vars:
\describe{
  \item{text}{text from interview}
   \item{speaker}{speaker identity}
  ...
}
}
\usage{
MaronGross_2013
}
\description{
Text and talker information delineated, raw transcript, multiple lines per talker
}
\keyword{datasets}
