% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MaryLittleLamb_Test}
\alias{MaryLittleLamb_Test}
\title{Text and talker information delineated, raw transcript, 100 turns}
\format{
## "Dyad_Sample"
A data.frame with 100 observations, 2 vars:
\describe{
  \item{text}{text from simulated interview}
   \item{speaker}{speaker identity}
  ...
}
}
\usage{
MaryLittleLamb_Test
}
\description{
Text and talker information delineated, raw transcript, 100 turns
}
\keyword{datasets}
