% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NurseryRhymes}
\alias{NurseryRhymes}
\title{Text and talker information delineated, 3 separate nursery rhymes, good for computing analytics and word counts}
\format{
## "NurseryRhymes"
A data.frame with 100 observations, 2 vars:
\describe{
  \item{Event_ID}{factor 3 different simulated conversations}
   \item{Participant_ID}{fictional speaker names, 2 each conversation}
   \item{Text_Raw}{simulated language production, actually looped phrases from nursery rhymes}
  ...
}
}
\usage{
NurseryRhymes
}
\description{
Text and talker information delineated, 3 separate nursery rhymes, good for computing analytics and word counts
}
\keyword{datasets}
