% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NurseryRhymes_Prepped}
\alias{NurseryRhymes_Prepped}
\title{Text and talker information delineated, 3 separate nursery rhymes, good for computing analytics and word counts}
\format{
## "NurseryRhymes_Prepped"
A data.frame with 1507 x 7 observations, 5 vars:
\describe{
  \item{Event_ID}{factor 3 different simulated conversations}
   \item{Participant_ID}{fictional speaker names, 2 each conversation}
   \item{Exchange_Count}{sequential numbering of exchanges by conversation, 1 exchange = 2 turns}
   \item{Turn_Count}{sequential numbering of turns by conversation}
   \item{Text_Clean}{content words}
    \item{emo_anger}{raw value of anger salience yoked to each word}
  ...
}
}
\usage{
NurseryRhymes_Prepped
}
\description{
Text and talker information delineated, 3 separate nursery rhymes, good for computing analytics and word counts
}
\keyword{datasets}
