% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dyads.R
\name{prep_dyads}
\alias{prep_dyads}
\title{prep_dyads}
\usage{
prep_dyads(
  read_ts_df,
  lemmatize = TRUE,
  omit_stops = TRUE,
  which_stoplist = "Temple_stops25"
)
}
\arguments{
\item{read_ts_df}{data frame produced from the read_dyads() function}

\item{lemmatize}{logical, should words be lemmatized (switched to base morphological form), default is TRUE}

\item{omit_stops}{remove stopwords, default TRUE}

\item{which_stoplist}{user specifies stopword removal method with options including "none", "SMART", "MIT_stops", "CA_OriginalStops", or "Temple_Stopwords25". "Temple_Stopwords25 is the default list}
}
\value{
dataframe with cleaned text data, formatted with one word per row
}
\description{
Cleans, vectorizes and appends lexical norms to all content words in a language corpus. User guides options for stopword removal and lemmatization. User selects up to three psycholinguistic dimensions to yoke norms on each content word in the transcript.
}
