% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dyads.R
\name{summarize_dyads}
\alias{summarize_dyads}
\title{summarize_dyads}
\usage{
summarize_dyads(aligned_ts_df, resample = TRUE, threshold = "min")
}
\value{
dataframe
}
\description{
appends AUC and Spearman Rank Correlation indices to each dyad (event_id) using a resampling algoirthm that defaults to the minimum number of exchanges across all documents entered
}
