% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dyads_auc.R
\name{summarize_dyads_auc}
\alias{summarize_dyads_auc}
\title{summarize_dyads_auc}
\usage{
summarize_dyads_auc(aligned_ts_df, resample = TRUE)
}
\arguments{
\item{dataframe}{produced in the align_dyads function}
}
\value{
A data frame associating each transcript with the dAUC calculated for each dimension
}
\description{
Calculates the area under the curve of the absolute difference time series between interlocutor time series. The length of the difference time series can be standardized the shortest number of exchanges present in the group using an internally defined resampling function, called with resample = TRUE. Area under the curve become less reliable for dyads under 30 exchanges.
}
