% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dyads_covar.R
\name{summarize_dyads_covar}
\alias{summarize_dyads_covar}
\title{summarize_dyads_covar}
\usage{
summarize_dyads_covar(aligned_ts_df, lags = c(-3, -2, -1, 0, 1, 2, 3))
}
\arguments{
\item{aligned_ts_df}{Dataframe produced in the align_dyads function}

\item{lags}{A vector of signed integers specifying lags that should be included for Pearson correlation. Negative integers are used for leads.}
}
\value{
A data frame containing the Spearman and Pearson correlation for each transcript. Results are organized as column and data is pivoted longer by dimension for readability.
}
\description{
Calculates Spearman rank correlation and lagged Pearson correlation between interlocutor time series for each dimension and transcript.
}
