% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_dyads.R
\name{align_dyads}
\alias{align_dyads}
\title{align_dyads}
\usage{
align_dyads(clean_ts_df)
}
\arguments{
\item{clean_ts_df}{cleaned and formatted dataframe ported from the clean_dyads() step}
}
\description{
Yokes user-specified semantic, affective, and phonological values to each word in a cleaned language transcript. Prepares a dataframe aligned by exchange and turn across Participant_IDs.
}
