% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_dyads.R
\name{align_dyads}
\alias{align_dyads}
\title{align_dyads}
\usage{
align_dyads(clean_ts_df)
}
\arguments{
\item{clean_ts_df}{a dataframe cleaned and formatted during the read_dyads() function}
}
\value{
a dataframe one-word-per-row format with variables of interest appended
}
\description{
Yokes user-specified semantic, affective, and phonological values to each word in a cleaned language transcript. Prepares a dataframe aligned by exchange and turn across Participant_IDs.
}
