% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dyads.R
\name{clean_dyads}
\alias{clean_dyads}
\title{clean_dyads}
\usage{
clean_dyads(read_ts_df, lemmatize = TRUE)
}
\arguments{
\item{dataframe}{produced from the read_dyads() function}
}
\value{
dataframe with stopwords omitted, lemmatized words one per row
}
\description{
Cleans and formats language transcripts from the read stage. Removes non-alphabetic characters and stopwords. Language transcripts can be lemmatized by calling lemmatize = TRUE. Vectorizes each utterance and reports the total word count and mean word length by interlocutor in each dyad. Also reports the number of words in each turn.
}
