% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dyads.R
\name{clean_dyads}
\alias{clean_dyads}
\title{clean_dyads}
\usage{
clean_dyads(
  read_ts_df,
  lemmatize = TRUE,
  omit_stops = TRUE,
  which_stoplist = "Temple_stops25"
)
}
\arguments{
\item{read_ts_df}{data frame produced from the read_dyads() function}

\item{lemmatize}{logical, should words be lemmatized (switched to base morphological form), default is TRUE}

\item{omit_stops}{remove stopwords, default TRUE}

\item{which_stoplist}{user specifies stopword removal method with options including "none", "SMART", "MIT_stops", "CA_OriginalStops", or "Temple_Stopwords25". "Temple_Stopwords25 is the default list}
}
\value{
dataframe with cleaned text data, formatted with one word per row
}
\description{
Cleans and formats language transcripts from the read stage. Removes non-alphabetic characters and stopwords. Language transcripts can be lemmatized by calling lemmatize = TRUE. Vectorizes each utterance and reports the total word count and mean word length by interlocutor in each dyad. Also reports the number of words in each turn.
}
