% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConversationAlign-package.R
\docType{package}
\name{ConversationAlign-package}
\alias{ConversationAlign}
\alias{ConversationAlign-package}
\title{ConversationAlign: Process Text and Compute Linguistic Alignment in Conversation Transcripts}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

`ConversationAlign` imports conversation transcripts into R, concatenates them into a single dataframe appending event identifiers, cleans and formats the text, then yokes user-specified psycholinguistic database values to each word. `ConversationAlign` then computes alignment indices between two interlocutors across each transcript for >40 possible semantic, lexical, and affective dimensions. In addition to alignment, `ConversationAlign` also produces a table of analytics (e.g., token count, type-token-ratio) in a summary table describing your particular text corpus.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Reilly-ConceptsCognitionLab/ConversationAlign}
}

}
\author{
\strong{Maintainer}: Jamie Reilly \email{jamie_reilly@temple.edu} (\href{https://orcid.org/0000-0002-0891-438X}{ORCID})

Authors:
\itemize{
  \item Virginia Ulichney
  \item Ben Sacks
}

Other contributors:
\itemize{
  \item Sarah Weinstein [contributor]
  \item Chelsea Helion [contributor]
  \item Gus Cooney [contributor]
}

}
\keyword{internal}
