% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_1file.R
\name{read_1file}
\alias{read_1file}
\title{read_1file}
\usage{
read_1file(my_dat)
}
\arguments{
\item{my_dat}{conversation transcript in csv or txt format}
}
\value{
a dataframe formatted with 'Event_ID', "Participant_ID", "RawText" -- ready for clean_dyads()
}
\description{
Reads pre-formatted dyadic (2 interlocutor) conversation transcript already imported into your R environment.
}
