% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dyads.R
\name{read_dyads}
\alias{read_dyads}
\title{read_dyads}
\usage{
read_dyads(folder_name = "my_transcripts")
}
\arguments{
\item{folder_name}{user can specify a folder name and directory for where their language transcripts will be read from, default is 'my-transcripts' in root}
}
\description{
Reads pre-formatted conversation transcripts from txt or csv on user's machine
}
