% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dyads.R
\name{read_dyads}
\alias{read_dyads}
\title{read_dyads}
\usage{
read_dyads(folder_name = "my_transcripts")
}
\arguments{
\item{folder_name}{folder of conversation transcripts in csv or txt format}
}
\value{
a concatenated dataframe with each language transcript saved as a separate 'event_id'; these are split into separate lists for discrete operations in later steps
}
\description{
Reads pre-formatted conversation transcripts from txt or csv on user's machine; user supplies directory path (e.g., "my_transcripts") to local folder as argument to function call
}
