% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dyads.R
\name{read_dyads}
\alias{read_dyads}
\title{read_dyads}
\usage{
read_dyads(folder_name = "my_transcripts")
}
\arguments{
\item{folder_name}{folder of conversation transcripts in csv or txt format}
}
\value{
a concatenated dataframe with each language transcript saved as a separate 'event_id'
}
\description{
Reads pre-formatted dyadic (2 interlocutor) conversation transcripts from your machine. Transcripts must be either csv or txt format. IF you are supplying a txt file, your transcript must be formatted as an otter.ai txt file export. Your options for using csv files are more flexible. ConversationAlign minimally requires a csv file with two columns, denoting interlocutor and text. Each separate conversation transcript should be saved as a separate file. ConversationAlign will use the file names as a document ID. Within the read dyads function, set the folder_name argument as the directory path to the local folder containing your transcripts on your machine (e.g., "my_transcripts"). Please see our github page for examples of properly formatted transcripts: https://github.com/Reilly-ConceptsCognitionLab/ConversationAlign
}
