% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dyads.R, R/summarize_dyads_new.R
\name{summarize_dyads}
\alias{summarize_dyads}
\title{summarize_dyads}
\usage{
summarize_dyads(df_prep, additional_lags = NULL)

summarize_dyads(df_prep, additional_lags = NULL)
}
\arguments{
\item{df_prep}{produced in the align_dyads function}

\item{additional_lags}{integer vector, should any lags be added in addition to -2, 0, 2}

\item{aligned_ts_df}{produced in the align_dyads function}
}
\description{
Calculates and appends 3 measures for quantifying alignment. Appends the mean score for each dimension by turn. Calculates and Spearman's rank correlation between interlocutor time series and appends by transcript. Calculates the area under the curve of the absolute difference time series between interlocutor time series. The length of the difference time series can be standardized the shortest number of exchanges present in the group using an internally defined resampling function, called with resample = TRUE. Spearman's rank correlation and area under the curve become less reliable for dyads under 30 exchanges.

Calculates and appends 3 measures for quantifying alignment. Appends the mean score for each dimension by turn. Calculates and Spearman's rank correlation between interlocutor time series and appends by transcript. Calculates the area under the curve of the absolute difference time series between interlocutor time series. The length of the difference time series can be standardized the shortest number of exchanges present in the group using an internally defined resampling function, called with resample = TRUE. Spearman's rank correlation and area under the curve become less reliable for dyads under 30 exchanges.
}
