% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dyads_slope.R
\name{summarize_dyads_slope}
\alias{summarize_dyads_slope}
\title{summarize_dyads_slope}
\usage{
summarize_dyads_slope(aligned_ts_df)
}
\arguments{
\item{aligned_ts_df}{data frame produced in the align_dyads function}
}
\value{
A data frame containing the intercept and slope of a linear regression of difference and interlocutor time series for each dimension
}
\description{
Calculates the intercept and slope of a simple linear regression for each interlocutor and the difference time series between them.
}
